+++ e2e-tests/cypress/e2e/datasets.cy.js
@@ -15,8 +15,20 @@ describe('Datasets page', () => {
             .should('have.length.greaterThan', 2);
     });
 
+    it('opens filters', () => {
+        cy.checkOpenFilters(0);
+    });
+
+    it('can filter datasets by owner', () => {
+        cy.checkFilterAssetsBy('owner');
+    });
+
     it('navigates to the dedicated dataset page', () => {
-        cy.get('tbody[data-cy=loaded]').get('tr').eq(2).click({ force: true });
+        cy.get('tbody[data-cy=loaded]').get('tr').eq(1).click({ force: true });
         cy.url().should('match', /datasets\/.{36}/);
+    });
+
+    it('searches dataset with a key', () => {
+        cy.checkSearchByKey('datasets');
     });
 });