+++ e2e-tests/cypress/support/commands.js
@@ -24,6 +24,15 @@
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
 
+// get the data-cy attribute
+Cypress.Commands.add('getDataCy', (input) => {
+    return cy.get(`[data-cy='${input}']`);
+});
+
+/**
+ * Login to the app using credentials in cypress env,
+ * then check if cookies have been set correctly.
+ */
 Cypress.Commands.add('login', () => {
     cy.session(
         'login',
@@ -40,6 +49,96 @@ Cypress.Commands.add('login', () => {
                 cy.getCookie('signature').should('not.be.null');
                 cy.getCookie('header.payload').should('not.be.null');
             },
+            cacheAcrossSpecs: true,
         }
     );
+});
+
+/**
+ * Check the item count to see if there is several pages.
+ * If there is, test to go back & forth between page 1 & 2.
+ * First using next & previous buttons,then using number buttons.
+ */
+Cypress.Commands.add('paginationTest', () => {
+    cy.getDataCy('items-count').then(($count) => {
+        const count = parseInt($count.text());
+        if (count > Cypress.env('DEFAULT_PAGE_SIZE')) {
+            cy.getDataCy('active-page').invoke('text').should('eq', '1');
+            cy.getDataCy('next-page').click();
+            cy.getDataCy('active-page').invoke('text').should('eq', '2');
+            cy.getDataCy('previous-page').click();
+            cy.getDataCy('active-page').invoke('text').should('eq', '1');
+            cy.getDataCy('second-page').click();
+            cy.getDataCy('active-page').invoke('text').should('eq', '2');
+        }
+    });
+});
+
+/**
+ * Check the opening and closing of the filter modal.
+ * First by clicking the add filter button,
+ * then by clicking the filter button inside table column header menus
+ */
+Cypress.Commands.add('checkOpenFilters', (position) => {
+    cy.getDataCy('filters-table').should('not.exist');
+    cy.getDataCy('add-filter').click();
+    cy.getDataCy('filters-table').should('exist');
+    cy.getDataCy('close-filters-modal').click();
+    cy.getDataCy('filters-table').should('not.be.visible');
+    cy.getDataCy('th-menu-button').eq(position).click();
+    cy.getDataCy('open-filters').first().click();
+    cy.getDataCy('filters-table').should('be.visible');
+});
+
+/**
+ * Check if an asset table can be filtered by a given filter.
+ * filter - expects the name of the filter to test as a string
+ */
+Cypress.Commands.add('checkFilterAssetsBy', (filter) => {
+    cy.getDataCy('add-filter').click();
+    cy.getDataCy(`${filter}-filters`).click();
+    cy.get('[data-cy="filter-checkbox"] > input')
+        .first()
+        .check({ force: true });
+    cy.getDataCy('filters-apply').click();
+    cy.getDataCy(`${filter}-filter-tag`).should('exist');
+});
+
+/**
+ * Check if a drawer is correctly opened when clicking an asset in its table.
+ */
+Cypress.Commands.add('checkOpenDrawer', () => {
+    cy.getDataCy('drawer').should('not.exist');
+    cy.get('tbody[data-cy=loaded]').get('tr').eq(1).click({ force: true });
+    cy.getDataCy('drawer').should('exist');
+});
+
+/**
+ * Check if a file of a given name has been correctly downloaded.
+ * filename - expects the filename as a string
+ */
+Cypress.Commands.add('checkDownloadedFile', (filename) => {
+    cy.readFile(`cypress/downloads/${filename}`, { timeout: 20000 });
+});
+
+Cypress.Commands.add('checkSearchByKey', (asset) => {
+    cy.get('tbody[data-cy=loaded]').should('exist');
+    cy.getDataCy('items-count').then(($count) => {
+        const count = parseInt($count.text());
+        if (count > Cypress.env('DEFAULT_PAGE_SIZE')) {
+            cy.get('[data-key]')
+                .first()
+                .invoke('data', 'key')
+                .then((key) => {
+                    cy.getDataCy('search-bar').type(key);
+                    cy.getDataCy('search-item').click();
+                    cy.url().should(
+                        'include',
+                        asset === 'compute_plans'
+                            ? `/${asset}/${key}/tasks`
+                            : `/${asset}/${key}`
+                    );
+                });
+        }
+    });
 });