+++ e2e-tests/cypress/support/e2e.js
@@ -27,4 +27,15 @@ beforeEach(() => {
             password: Cypress.env('PASSWORD'),
         }
     );
+
+    /**
+     * keep uncaught exceptions from failing tests
+     * uncomment this function to test part with expected error
+     **/
+    // cy.on('uncaught:exception', (e, runnable) => {
+    //     console.log('error', e);
+    //     console.log('runnable', runnable);
+    //     console.log('error', e.message);
+    //     return false;
+    // });
 });