+++ src/features/tableFilters/TableFilters.tsx
@@ -107,13 +107,21 @@ const TableFilters = ({ children }: TableFiltersProps): JSX.Element => {
                         size="sm"
                         backgroundColor="secondary"
                         leftIcon={<RiAddLine />}
+                        data-cy="add-filter"
                     >
                         Add Filter
                     </Button>
                 </PopoverTrigger>
-                <PopoverContent minWidth="690px" boxShadow="xl">
+                <PopoverContent
+                    minWidth="690px"
+                    boxShadow="xl"
+                    data-cy="filters-table"
+                >
                     <PopoverArrow />
-                    <PopoverCloseButton onClick={onPopoverClose} />
+                    <PopoverCloseButton
+                        onClick={onPopoverClose}
+                        data-cy="close-filters-modal"
+                    />
                     <PopoverBody padding="0" overflow="hidden">
                         <Tabs
                             variant="soft-rounded"
@@ -142,6 +150,7 @@ const TableFilters = ({ children }: TableFiltersProps): JSX.Element => {
                                                 ? initialFocusRef
                                                 : null
                                         }
+                                        data-cy={`${tab.title.toLowerCase()}-filters`}
                                     >
                                         {tab.title}
                                         {isActive(tab.field) && (
@@ -189,6 +198,7 @@ const TableFilters = ({ children }: TableFiltersProps): JSX.Element => {
                                 color="white"
                                 colorScheme="primary"
                                 onClick={onApply}
+                                data-cy="filters-apply"
                             >
                                 Apply
                             </Button>